Agenda Contacto - Opcion eliminar contacto

+++ agenda_contactos/constantes.c
@@ -1,4 +1,4 @@
 # include "constantes.h"
 
 Contacto **GLOBAL_CONTACTOS = NULL;
-unsigned int GLOBAL_LEN_CONTACTOS = 0;
+size_t GLOBAL_LEN_CONTACTOS = 0;
